Move markdown widget builder to src
